refactor: classify the logic for updating data by keyMap or pasting
