+++ packages/toast-ui.grid/src/dispatch/data.ts
@@ -7,13 +7,10 @@ import {
   RemoveTargetRows,
 } from '@t/store/data';
 import { Store } from '@t/store';
-import { SelectionRange } from '@t/store/selection';
 import { ColumnInfo } from '@t/store/column';
 import { OptRow, OptAppendRow, OptRemoveRow, ResetOptions } from '@t/options';
-import { copyDataToRange, getRangeToPaste } from '../query/clipboard';
 import {
   findProp,
-  mapProp,
   removeArrayItem,
   includes,
   isEmpty,
@@ -23,7 +20,7 @@ import {
 } from '../helper/common';
 import { createViewRow, createData, setRowRelationListItems, createRawRow } from '../store/data';
 import { notify, isObservable, batchedInvokeObserver } from '../helper/observable';
-import { changeSelectionRange, initSelection } from './selection';
+import { initSelection } from './selection';
 import { getEventBus } from '../event/eventBus';
 import GridEvent from '../event/gridEvent';
 import { getDataManager, getDataProvider } from '../instance';
@@ -42,7 +39,6 @@ import {
 import {
   findIndexByRowKey,
   findRowByRowKey,
-  isEditableCell,
   getRowHeight,
   getLoadingState,
   getAddedClassName,
@@ -60,7 +56,6 @@ import {
   updateAllSummaryValues,
 } from './summary';
 import { initFilter, resetFilterState } from './filter';
-import { getSelectionRange } from '../query/selection';
 import { initScrollPosition } from './viewport';
 import { isRowHeader } from '../helper/column';
 import { updatePageOptions, updatePageWhenRemovingRow, resetPageState } from './pagination';
@@ -160,7 +155,7 @@ export function setValue(
   /**
    * Occurs before one or more cells is changed
    * @event Grid#beforeChange
-   * @property {number | string} rowKey - rowKey of the unchecked row
+   * @property {string} changeType - The type of change('paste', 'delete', 'cell')
    * @property {Array.<object>} changes - rowKey, column name, original values and next values before changing the values
    * @property {Grid} instance - Current grid instance
    */
@@ -211,7 +206,7 @@ export function setValue(
   /**
    * Occurs after one or more cells is changed
    * @event Grid#afterChange
-   * @property {number | string} rowKey - rowKey of the unchecked row
+   * @property {string} changeType - The type of change('paste', 'delete', 'cell')
    * @property {Array.<object>} changes - rowKey, column name, previous values and changed values after changing the values
    * @property {Grid} instance - Current grid instance
    */
@@ -358,127 +353,6 @@ export function uncheckAll(store: Store, allPage?: boolean) {
    * @property {Grid} instance - Current grid instance
    */
   eventBus.trigger('uncheckAll', gridEvent);
-}
-
-function createChangeInfo(
-  store: Store,
-  row: Row,
-  columnName: string,
-  pastingValue: CellValue,
-  index: number
-) {
-  const { id, column } = store;
-  const { rowKey } = row;
-  const prevChange = { rowKey, columnName, value: row[columnName], nextValue: pastingValue };
-  const nextChange = { rowKey, columnName, prevValue: row[columnName], value: pastingValue };
-  const changeValue = () => {
-    const { value, nextValue } = prevChange;
-    replaceColumnUniqueInfoMap(id, column, {
-      rowKey,
-      columnName,
-      prevValue: value,
-      value: nextValue,
-    });
-    nextChange.value = nextValue;
-    row[columnName] = nextValue;
-
-    return index;
-  };
-
-  return { prevChange, nextChange, changeValue };
-}
-
-function applyPasteDataToRawData(
-  store: Store,
-  pasteData: string[][],
-  indexToPaste: SelectionRange
-) {
-  const { data, column, id } = store;
-  const eventBus = getEventBus(id);
-  const { filteredRawData, filteredViewData } = data;
-  const { visibleColumnsWithRowHeader } = column;
-  const {
-    row: [startRowIndex, endRowIndex],
-    column: [startColumnIndex, endColumnIndex],
-  } = indexToPaste;
-
-  const columnNames = mapProp('name', visibleColumnsWithRowHeader);
-  const changeValueFns = [];
-  const prevChanges = [];
-  const nextChanges = [];
-
-  for (let rowIndex = 0; rowIndex + startRowIndex <= endRowIndex; rowIndex += 1) {
-    const rawRowIndex = rowIndex + startRowIndex;
-    for (let columnIndex = 0; columnIndex + startColumnIndex <= endColumnIndex; columnIndex += 1) {
-      const name = columnNames[columnIndex + startColumnIndex];
-      if (filteredViewData.length && isEditableCell(data, column, rawRowIndex, name)) {
-        const targetRow = filteredRawData[rawRowIndex];
-        const { prevChange, nextChange, changeValue } = createChangeInfo(
-          store,
-          targetRow,
-          name,
-          pasteData[rowIndex][columnIndex],
-          rawRowIndex
-        );
-        prevChanges.push(prevChange);
-        nextChanges.push(nextChange);
-        changeValueFns.push(changeValue);
-      }
-    }
-  }
-  let gridEvent = new GridEvent({
-    changeType: 'paste',
-    changes: prevChanges,
-  });
-
-  /**
-   * Occurs before one or more cells is changed
-   * @event Grid#beforeChange
-   * @property {number | string} rowKey - rowKey of the unchecked row
-   * @property {Array.<object>} changes - rowKey, column name, original values and next values before changing the values
-   * @property {Grid} instance - Current grid instance
-   */
-  eventBus.trigger('beforeChange', gridEvent);
-
-  if (!gridEvent.isStopped()) {
-    let index: number | null = null;
-    changeValueFns.forEach((changeValue) => {
-      const targetRowIndex = changeValue();
-      if (index !== targetRowIndex) {
-        index = targetRowIndex;
-        getDataManager(id).push('UPDATE', filteredRawData[index]);
-      }
-    });
-    forceValidateUniquenessOfColumns(data.rawData, column);
-
-    gridEvent = new GridEvent({
-      changeType: 'paste',
-      changes: nextChanges,
-    });
-
-    /**
-     * Occurs after one or more cells is changed
-     * @event Grid#afterChange
-     * @property {number | string} rowKey - rowKey of the unchecked row
-     * @property {Array.<object>} changes - rowKey, column name, previous values and changed values after changing the values
-     * @property {Grid} instance - Current grid instance
-     */
-    eventBus.trigger('afterChange', gridEvent);
-  }
-}
-
-export function paste(store: Store, pasteData: string[][]) {
-  const { selection, id, data } = store;
-  const { pageOptions } = data;
-  const { originalRange } = selection;
-
-  if (originalRange) {
-    pasteData = copyDataToRange(originalRange, pasteData);
-  }
-
-  const rangeToPaste = getRangeToPaste(store, pasteData);
-  applyPasteDataToRawData(store, pasteData, rangeToPaste);
-  changeSelectionRange(selection, getSelectionRange(rangeToPaste, pageOptions), id);
 }
 
 export function setDisabledAllCheckbox({ data }: Store) {
